// src/modules/main/home/store/types.ts
export enum CREATE_POST {
  REQUEST = 'home/CREATE_POST_REQUEST',
  SUCCESS = 'home/CREATE_POST_SUCCESS',
  FAILURE = 'home/CREATE_POST_FAILURE',
}

export enum GET_POSTS {
  REQUEST = 'home/GET_POSTS_REQUEST',
  SUCCESS = 'home/GET_POSTS_SUCCESS',
  FAILURE = 'home/GET_POSTS_FAILURE',
  ADD = 'home/GET_POSTS_ADD',
  REMOVE = 'home/GET_POSTS_REMOVE',
}

export enum GET_COMMENTS_FROM_POST {
  REQUEST = 'home/GET_COMMENTS_FROM_POST_REQUEST',
  SUCCESS = 'home/GET_COMMENTS_FROM_POST_SUCCESS',
  FAILURE = 'home/GET_COMMENTS_FROM_POST_FAILURE',
}

export enum DELETE_POST {
  REQUEST = 'home/DELETE_POST_REQUEST',
  SUCCESS = 'home/DELETE_POST_SUCCESS',
  FAILURE = 'home/DELETE_POST_FAILURE',
}

export enum CREATE_COMMENT {
  REQUEST = 'home/CREATE_COMMENT_REQUEST',
  SUCCESS = 'home/CREATE_COMMENT_SUCCESS',
  FAILURE = 'home/CREATE_COMMENT_FAILURE',
}